Move recipe classes to separate files. Highlight cell background in gold when editting a cell
