+++ StonehearthEditor/Recipes/RowMetadata.cs
@@ -1,0 +1,13 @@
+﻿
+namespace StonehearthEditor
+{
+
+    internal class RowMetadata
+    {
+        public JsonFileData RecipeList { get; set; }
+
+        public JsonFileData Recipe { get; set; }
+
+        public JsonFileData Item { get; set; }
+    }
+}
